requirements/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.1
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.1.0
idna==2.9
importlib-metadata==5.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
jwt==1.0.0
Markdown==3.4.1
MarkupSafe==1.1.1
marshmallow==3.5.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pipenv==2018.11.26
psycopg2==2.8.4
pycparser==2.20
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.4.3
tzdata==2022.4
tzlocal==2.0.0
urllib3==1.25.8
validators==0.20.0
virtualenv==16.7.8
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wrapt==1.14.1
zipp==3.8.1
